+++ foxes_opt/problems/layout/regular_layout.py
@@ -8,7 +8,7 @@ import foxes.variables as FV
 import foxes.constants as FC
 
 
-def _calc_func(valid, P, ct, st_sel):
+def _calc_func(valid, P, ct, algo, mdata, fdata, st_sel):
     """helper function for Calculator turbine model"""
     return (valid, P * valid, ct * valid)
 
@@ -29,8 +29,8 @@ class RegularLayoutOptProblem(FarmVarsProblem):
 
     SPACING_X = "spacing_x"
     SPACING_Y = "spacing_y"
-    OFFSET_X = "offset_X"
-    OFFSET_Y = "offset_Y"
+    OFFSET_X = "offset_x"
+    OFFSET_Y = "offset_y"
     ANGLE = "angle"
 
     def __init__(
@@ -88,7 +88,7 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         pmax = b.p_max()
         pmin = b.p_min()
         self._pmin = pmin
-        self._xy0 = 0.5 * (pmin + pmax)
+        self._xy0 = pmin#0.5 * (pmin + pmax)
         self._halfspan = (pmax - pmin) / 2
         self._halflen = np.linalg.norm(self._halfspan)
         self.max_spacing = 2 * (self._halflen + self.min_spacing)
@@ -116,7 +116,6 @@ class RegularLayoutOptProblem(FarmVarsProblem):
                 self.farm.turbines[-1].name = f"T{ti}"
         elif self.farm.n_turbines > self._nturb:
             self.farm.turbines = self.farm.turbines[: self._nturb]
-        self.algo.n_turbines = self._nturb
         self.algo.update_n_turbines()
 
         super().initialize(
@@ -168,13 +167,13 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             Minimal float values, shape: (n_vars_float,)
 
         """
-        return [
+        return np.array([
             self.min_spacing,
             self.min_spacing,
-            -self._halfspan[0] - self.min_spacing,
-            -self._halfspan[1] - self.min_spacing,
+            - self.max_spacing,
+            - self.max_spacing,
             0.0,
-        ]
+        ], dtype=config.dtype_double)
 
     def max_values_float(self):
         """
@@ -188,13 +187,13 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             Maximal float values, shape: (n_vars_float,)
 
         """
-        return [
+        return np.array([
             self.max_spacing,
             self.max_spacing,
-            self._halfspan[0] + self.min_spacing,
-            self._halfspan[1] + self.min_spacing,
+            self.max_spacing,
+            self.max_spacing,
             90.0,
-        ]
+        ], dtype=config.dtype_double)
 
     def opt2farm_vars_individual(self, vars_int, vars_float):
         """
@@ -228,21 +227,21 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         nay = np.cross(np.array([0.0, 0.0, 1.0], dtype=config.dtype_double), nax)
         x0 = self._xy0 + np.array([ox, oy], dtype=config.dtype_double)
 
-        pts = np.zeros((n_states, nx, ny, 2), dtype=config.dtype_double)
+        pts = np.zeros((nx, ny, 2), dtype=config.dtype_double)
         pts[:] = (
-            x0[None, None, None, :]
-            + np.arange(nx)[None, :, None, None] * dx * nax[None, None, None, :2]
-            + np.arange(ny)[None, None, :, None] * dy * nay[None, None, None, :2]
+            x0[None, None, :]
+            + np.arange(nx)[:, None, None] * dx * nax[None, None, :2]
+            + np.arange(ny)[None, :, None] * dy * nay[None, None, :2]
         )
 
-        pts = pts.reshape(n_states, nx * ny, 2)
-        valid = self.farm.boundary.points_inside(pts.reshape(n_states * nx * ny, 2))
+        pts = pts.reshape(nx * ny, 2)
+        valid = self.farm.boundary.points_inside(pts)
 
-        farm_vars = {
-            FV.X: pts[:, :, 0],
-            FV.Y: pts[:, :, 1],
-            FC.VALID: valid.reshape(n_states, nx * ny).astype(config.dtype_double),
-        }
+        farm_vars = {}
+        for v, d in zip([FV.X, FV.Y, FC.VALID], [pts[:, 0], pts[:, 1], valid]):
+            a = np.zeros((n_states, nx * ny), dtype=config.dtype_double)
+            a[:] = d[None, :]
+            farm_vars[v] = a
 
         return farm_vars
 
@@ -286,34 +285,32 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         naz[..., 2] = 1
         nay = np.cross(naz, nax)
 
-        pts = np.zeros((n_pop, n_states, nx, ny, 2), dtype=config.dtype_double)
-        pts[:] = self._xy0[None, None, None, None, :]
-        pts[..., 0] += ox[:, None, None, None]
-        pts[..., 1] += oy[:, None, None, None]
+        pts = np.zeros((n_pop, nx, ny, 2), dtype=config.dtype_double)
+        pts[:] = self._xy0[None, None, None, :]
+        pts[..., 0] += ox[:, None, None]
+        pts[..., 1] += oy[:, None, None]
         pts[:] += (
-            np.arange(nx)[None, None, :, None, None]
-            * dx[:, None, None, None, None]
-            * nax[:, None, None, None, :2]
-            + np.arange(ny)[None, None, None, :, None]
-            * dy[:, None, None, None, None]
-            * nay[:, None, None, None, :2]
+            np.arange(nx)[None, :, None, None]
+            * dx[:, None, None, None]
+            * nax[:, None, None, :2]
+            + np.arange(ny)[None, None, :, None]
+            * dy[:, None, None, None]
+            * nay[:, None, None, :2]
         )
-
-        qts = np.zeros((n_pop, n_states, n_turbines, 2))
-        qts[:, :N] = pts.reshape(n_pop, n_states, N, 2)
+    
+        qts = np.zeros((n_pop, n_turbines, 2))
+        qts[:, :N] = pts.reshape(n_pop, N, 2)
         del pts
 
         valid = self.farm.boundary.points_inside(
-            qts.reshape(n_pop * n_states * n_turbines, 2)
-        )
+            qts.reshape(n_pop * n_turbines, 2)
+        ).reshape(n_pop, n_turbines)
 
-        farm_vars = {
-            FV.X: qts[:, :, :, 0],
-            FV.Y: qts[:, :, :, 1],
-            FC.VALID: valid.reshape(n_pop, n_states, n_turbines).astype(
-                config.dtype_double
-            ),
-        }
+        farm_vars = {}
+        for v, d in zip([FV.X, FV.Y, FC.VALID], [qts[:, :, 0], qts[:, :, 1], valid]):
+            a = np.zeros((n_pop, n_states, n_turbines), dtype=config.dtype_double)
+            a[:] = d[:, None, :]
+            farm_vars[v] = a
 
         return farm_vars